Добавление заданий к третьему уроку

+++ les2/dz2_5.py
@@ -13,7 +13,7 @@ while True:
     rate = (input("Введите новый элемент рейтинга в виде натурального числа, для остановки ввода введите brake: "))
     if rate == "brake":
         break
-    list_1.append(int(rate))
+    list_1.append(float(rate))
     list_1.sort()
     print(list_1[::-1])
